simg2img script
